Revert "game-data: decoration can be int16_t."
This reverts commit af1f2db9077da2e0a54a1f940ded78104445a606.

+++ src/game-data.h
@@ -88,7 +88,7 @@ struct ItemConfig {
 };
 
 struct TileConfig {
-    int16_t decoration;
+    int decoration;
     uint8_t type;
     uint8_t blocking;
 };
